# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CreatorHub API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /profile:
    get:
      operationId: getProfile
      tags: [profile]
      summary: Get current user's profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
        "404":
          description: Profile not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    patch:
      operationId: updateProfile
      tags: [profile]
      summary: Update current user's profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
  /profile/public/{username}:
    get:
      operationId: getPublicProfile
      tags: [profile]
      summary: Get public profile by username
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Public profile with links and products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicProfile"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  # ─── LINKS ──────────────────────────────────────────────────────────────────
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products for current user
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a new digital product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── POSTS / CONTENT PLANNER ─────────────────────────────────────────────────
...
  /subscription/upgrade:
    post:
      operationId: upgradeSubscription
      tags: [subscription]
      summary: Upgrade or change subscription plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpgradeSubscriptionInput"
      responses:
        "200":
          description: Updated subscription
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriptionInfo"

  # ─── WALLET ───────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Profile:
      type: object
      properties:
        id:
          type: integer
        clerkId:
          type: string
        name:
          type: string
        username:
          type: string
        email:
          type: string
        bio:
          type: ["string", "null"]
        profileImage:
          type: ["string", "null"]
        whatsappNumber:
          type: ["string", "null"]
        theme:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - clerkId
        - name
        - username
        - email
        - theme
        - createdAt
    ProfileUpdate:
      type: object
      properties:
        name:
          type: string
        username:
          type: string
        bio:
          type: string
        profileImage:
          type: string
        whatsappNumber:
          type: string
        theme:
          type: string
    PublicProfile:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        username:
          type: string
        bio:
          type: ["string", "null"]
        profileImage:
          type: ["string", "null"]
        whatsappNumber:
          type: ["string", "null"]
        theme:
          type: string
        links:
          type: array
          items:
            $ref: "#/components/schemas/Link"
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
      required:
        - id
        - name
        - username
        - theme
        - links
        - products

    # ─── LINKS ──────────────────────────────────────────────────────────────
    Link:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        title:
          type: string
        url:
          type: string
        icon:
          type: ["string", "null"]
        clicks:
          type: integer
        sortOrder:
          type: integer
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - title
        - url
        - clicks
        - sortOrder
        - isActive
        - createdAt
...
    Product:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        currency:
          type: string
        imageUrl:
          type: ["string", "null"]
        fileUrl:
          type: ["string", "null"]
        salesCount:
          type: integer
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - name
        - price
        - currency
        - salesCount
        - isActive
        - createdAt
    ProductInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        price:
          type: number
          minimum: 0
        currency:
          type: string
        imageUrl:
          type: string
        fileUrl:
          type: string
      required:
        - name
        - price
    ProductUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        currency:
          type: string
        imageUrl:
          type: string
        fileUrl:
          type: string
        isActive:
          type: boolean

    # ─── POSTS ────────────────────────────────────────────────────────────────
...
    SubscriptionInfo:
      type: object
      properties:
        id:
          type: integer
        plan:
          type: string
          enum: [free, pro, business]
        status:
          type: string
        currentPeriodStart:
          type: ["string", "null"]
          format: date-time
        currentPeriodEnd:
          type: ["string", "null"]
          format: date-time
        cancelAtPeriodEnd:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - plan
        - status
        - cancelAtPeriodEnd
        - createdAt
...
    UpgradeSubscriptionInput:
      type: object
      properties:
        plan:
          type: string
          enum: [free, pro, business]
      required:
        - plan

    # ─── WALLET ───────────────────────────────────────────────────────────────
...